refactor(lenses): updates to composables
